int转long

+++ src/element/pages/Page.java
@@ -10,16 +10,16 @@ import java.util.List;
  */
 public class Page {
     private final List<Record> list;
-    private final int pageId;
+    private final long pageId;
     private int recordNum;
 
-    public Page(int pageId, int recordNum) {
+    public Page(long pageId, int recordNum) {
         this.pageId = pageId;
         this.recordNum = recordNum;
         this.list = new ArrayList<> ();
     }
 
-    public int getPageId() {
+    public long getPageId() {
         return this.pageId;
     }
