+++ src/element/records/Record.java
@@ -8,10 +8,10 @@ import config.TableConfig;
 public class Record {
     private final String[] fields;
     private final int size;
-    private int pageId;
+    private long pageId;
     private int recordId;
 
-    public Record(String[] fields, int pageId, int recordId) {
+    public Record(String[] fields, long pageId, int recordId) {
         this.size = fields.length;
         this.fields = fields;
         this.pageId = pageId;
@@ -36,7 +36,7 @@ public class Record {
         return fields;
     }
 
-    public int getPageId() {
+    public long getPageId() {
         return pageId;
     }
