+++ src/operate/Load.java
@@ -12,8 +12,8 @@ import java.io.RandomAccessFile;
 import java.text.ParseException;
 
 public class Load {
-    private static int pageNum;//加载的数据流文件总页数
-    private static int pagePoint;//当前需要解析的页指针
+    private static long pageNum;//加载的数据流文件总页数
+    private static long pagePoint;//当前需要解析的页指针
 
     private int pageSize;
     private String dataFilePath;
@@ -24,7 +24,7 @@ public class Load {
         this.pageSize = pageSize;
         this.dataFilePath = TableConfig.PAGENAME + TableConfig.POINT + String.valueOf (pageSize);
         raf = new RandomAccessFile(dataFilePath, "r");
-        pageNum = (int) raf.length () / pageSize;
+        pageNum = raf.length () / pageSize;
         pagePoint = 0;
     }
 
@@ -54,13 +54,13 @@ public class Load {
         byte[] numBytes = new byte[4];
         raf.readFully(numBytes);//当前页记录总条数
         int recordNum = TypeUtil.bytesToInt (numBytes);
-        return new Page (pagePoint + 1, recordNum);
+        return new Page ( pagePoint + 1, recordNum);
     }
 
     //对传入的页进行解析，查找指定的记录
     public void query(Page page) throws IOException, ParseException {
         byte[] recordByte = new byte[TableConfig.RECORDLENGTH];
-        raf.seek ((long) (page.getPageId () - 1) * pageSize);
+        raf.seek ((page.getPageId () - 1) * pageSize);
         for (int i = 0; i < page.getRecordNum (); i++) {
             raf.read(recordByte);
             String[] records = RecordUtil.parseRecord(recordByte);
